Class Sheet fixes, more refactoring
