+++ module/sheets/actor/character-sheet.mjs
@@ -1,12 +1,12 @@
-import doGMRoll from '../helpers/adversary-roll.mjs'
-import { DAGGERHEART } from '../helpers/config.mjs'
+import doGMRoll from '../../helpers/adversary-roll.mjs'
+import { DAGGERHEART } from '../../helpers/config.mjs'
 import {
   onManageActiveEffect,
   prepareActiveEffectCategories,
-} from '../helpers/effects.mjs'
-import openChoiceMenu from '../helpers/item-choices.mjs'
-import doDHRoll from '../helpers/roll-macro.mjs'
-import { DaggerHeartHandSheet } from './hand-sheet.mjs'
+} from '../../helpers/effects.mjs'
+import openChoiceMenu from '../../helpers/item-choices.mjs'
+import doDHRoll from '../../helpers/roll-macro.mjs'
+import { DaggerHeartHandSheet } from '../hand-sheet.mjs'
 
 // tabs.bind(window);
 /**
