+++ templates/item/item-card-sheet.hbs
@@ -12,7 +12,7 @@
       </select>
     </div>
     <div class="w-100">
-      <label class="label strong" for="system.feature" class="resource-label">Feature</label>
+      <label class="label strong" for="system.feature" class="resource-label">Description</label>
       {{{editor system.feature target="system.feature" class="simple h150 white" button=true owner=owner
                 editable=editable}}}
     </div>
